requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
google-cloud-storage==2.18.2
google-cloud-documentai==2.27.0
google-cloud-aiplatform==1.65.0
pdfplumber==0.11.4
PyPDF2==3.0.1
python-docx==1.1.2
httpx==0.27.2
pydantic==2.8.2
pytest==8.3.2
fastapi>=0.70.0
uvicorn>=0.17.0